# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Candlekeep Core",
    "description": "FastAPI backend for RP platform with character cards and chat management",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Models",
        "description": "List model definitions with pagination and filtering",
        "operationId": "list_models_api_models_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-based)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-based)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Limit"
            },
            "description": "Items per page"
          },
          {
            "name": "name__ilike",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name  Ilike"
            }
          },
          {
            "name": "provider_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider Id"
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Enabled"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ModelListResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Create Model",
        "description": "Create a new model definition",
        "operationId": "create_model_api_models_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Model",
        "description": "Get model definition by ID.\nReturns detailed information including the embedded Model Family.",
        "operationId": "get_model_api_models__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "models"
        ],
        "summary": "Update Model",
        "description": "Update model definition",
        "operationId": "update_model_api_models__model_id__put",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "Delete Model",
        "description": "Delete model definition",
        "operationId": "delete_model_api_models__model_id__delete",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}/flags": {
      "patch": {
        "tags": [
          "models"
        ],
        "summary": "Update Model Flags",
        "description": "Toggle model enabled and OpenRouter routing flags",
        "operationId": "update_model_flags_api_models__model_id__flags_patch",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelFlagsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chats": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "List Chats",
        "description": "List chats with cursor-based pagination and filtering",
        "operationId": "list_chats_api_chats_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 8601 timestamp cursor for pagination",
              "title": "Cursor"
            },
            "description": "ISO 8601 timestamp cursor for pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          },
          {
            "name": "character_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Character Id"
            }
          },
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "name": "created_at__ge",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created At  Ge"
            }
          },
          {
            "name": "created_at__le",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created At  Le"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ChatResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Create Chat",
        "description": "Create a new chat",
        "operationId": "create_chat_api_chats_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{chat_id}": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "Get Chat",
        "description": "Get chat details by ID",
        "operationId": "get_chat_api_chats__chat_id__get",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "chats"
        ],
        "summary": "Update Chat",
        "description": "Update chat (e.g., change title or model)",
        "operationId": "update_chat_api_chats__chat_id__put",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chats"
        ],
        "summary": "Delete Chat",
        "description": "Delete chat",
        "operationId": "delete_chat_api_chats__chat_id__delete",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelCreate": {
        "properties": {
          "provider_id": {
            "type": "string",
            "maxLength": 12,
            "minLength": 1,
            "title": "Provider Id",
            "description": "Provider ID"
          },
          "model_identifier": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Model Identifier",
            "description": "Actual API model name"
          },
          "openrouter_identifier": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Openrouter Identifier",
            "description": "OpenRouter model name"
          },
          "use_openrouter": {
            "type": "boolean",
            "title": "Use Openrouter",
            "description": "Whether to route through OpenRouter",
            "default": false
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "User-friendly display name"
          },
          "model_family_id": {
            "type": "string",
            "maxLength": 12,
            "minLength": 1,
            "title": "Model Family Id",
            "description": "Link to model family"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50000
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt",
            "description": "Custom system prompt override"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters",
            "description": "All model parameters (temperature, max_tokens, etc.)"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "Whether model is available",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "provider_id",
          "model_identifier",
          "name",
          "model_family_id"
        ],
        "title": "ModelCreate",
        "description": "Schema for creating a new model definition"
      },
      "ModelDetailResponse": {
        "properties": {
          "provider_id": {
            "type": "string",
            "maxLength": 12,
            "minLength": 1,
            "title": "Provider Id",
            "description": "Provider ID"
          },
          "model_identifier": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Model Identifier",
            "description": "Actual API model name"
          },
          "openrouter_identifier": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Openrouter Identifier",
            "description": "OpenRouter model name"
          },
          "use_openrouter": {
            "type": "boolean",
            "title": "Use Openrouter",
            "description": "Whether to route through OpenRouter",
            "default": false
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "User-friendly display name"
          },
          "model_family_id": {
            "type": "string",
            "maxLength": 12,
            "minLength": 1,
            "title": "Model Family Id",
            "description": "Link to model family"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50000
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt",
            "description": "Custom system prompt override"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters",
            "description": "All model parameters (temperature, max_tokens, etc.)"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "Whether model is available",
            "default": true
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "can_use_openrouter": {
            "type": "boolean",
            "title": "Can Use Openrouter"
          },
          "active_identifier": {
            "type": "string",
            "title": "Active Identifier"
          },
          "provider_enabled": {
            "type": "boolean",
            "title": "Provider Enabled"
          },
          "model_family": {
            "$ref": "#/components/schemas/ModelFamilyResponse"
          }
        },
        "type": "object",
        "required": [
          "provider_id",
          "model_identifier",
          "name",
          "model_family_id",
          "id",
          "created_at",
          "updated_at",
          "can_use_openrouter",
          "active_identifier",
          "provider_enabled",
          "model_family"
        ],
        "title": "ModelDetailResponse",
        "description": "Schema for detailed model responses (includes embedded relationships)"
      },
...
      "ModelFlagsUpdate": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "use_openrouter": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Openrouter"
          }
        },
        "type": "object",
        "title": "ModelFlagsUpdate",
        "description": "Schema for updating model flags only"
      },
...
      "ModelResponse": {
        "properties": {
          "provider_id": {
            "type": "string",
            "maxLength": 12,
            "minLength": 1,
            "title": "Provider Id",
            "description": "Provider ID"
          },
          "model_identifier": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Model Identifier",
            "description": "Actual API model name"
          },
          "openrouter_identifier": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Openrouter Identifier",
            "description": "OpenRouter model name"
          },
          "use_openrouter": {
            "type": "boolean",
            "title": "Use Openrouter",
            "description": "Whether to route through OpenRouter",
            "default": false
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "User-friendly display name"
          },
          "model_family_id": {
            "type": "string",
            "maxLength": 12,
            "minLength": 1,
            "title": "Model Family Id",
            "description": "Link to model family"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50000
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt",
            "description": "Custom system prompt override"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters",
            "description": "All model parameters (temperature, max_tokens, etc.)"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "Whether model is available",
            "default": true
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "can_use_openrouter": {
            "type": "boolean",
            "title": "Can Use Openrouter"
          },
          "active_identifier": {
            "type": "string",
            "title": "Active Identifier"
          },
          "provider_enabled": {
            "type": "boolean",
            "title": "Provider Enabled"
          }
        },
        "type": "object",
        "required": [
          "provider_id",
          "model_identifier",
          "name",
          "model_family_id",
          "id",
          "created_at",
          "updated_at",
          "can_use_openrouter",
          "active_identifier",
          "provider_enabled"
        ],
        "title": "ModelResponse",
        "description": "Schema for detailed model responses"
      },
      "ModelUpdate": {
        "properties": {
          "provider_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 12,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          },
          "model_identifier": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Identifier"
          },
          "openrouter_identifier": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Openrouter Identifier"
          },
          "use_openrouter": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Openrouter"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "model_family_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 12,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Family Id"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50000
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          },
          "parameters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameters"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "ModelUpdate",
        "description": "Schema for updating a model definition"
      },
...
      "PaginatedResponse_ChatResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChatResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "PaginatedResponse[ChatResponse]"
      },
...
      "PaginatedResponse_ModelListResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ModelListResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "PaginatedResponse[ModelListResponse]"
      },
...
    }
  }
}